+++ internal/pkg/context.go
@@ -109,7 +109,7 @@ func (c *CallCtxt) Int8(i int) int8   { return int8(c.intValue(i, 8, "int8")) }
 func (c *CallCtxt) Int16(i int) int16 { return int16(c.intValue(i, 16, "int16")) }
 func (c *CallCtxt) Int32(i int) int32 { return int32(c.intValue(i, 32, "int32")) }
 func (c *CallCtxt) Rune(i int) rune   { return rune(c.intValue(i, 32, "rune")) }
-func (c *CallCtxt) Int64(i int) int64 { return int64(c.intValue(i, 64, "int64")) }
+func (c *CallCtxt) Int64(i int) int64 { return c.intValue(i, 64, "int64") }
 
 func (c *CallCtxt) intValue(i, bits int, typ string) int64 {
 	arg := c.args[i]
@@ -132,7 +132,7 @@ func (c *CallCtxt) Uint8(i int) uint8   { return uint8(c.uintValue(i, 8, "uint8"
 func (c *CallCtxt) Byte(i int) uint8    { return byte(c.uintValue(i, 8, "byte")) }
 func (c *CallCtxt) Uint16(i int) uint16 { return uint16(c.uintValue(i, 16, "uint16")) }
 func (c *CallCtxt) Uint32(i int) uint32 { return uint32(c.uintValue(i, 32, "uint32")) }
-func (c *CallCtxt) Uint64(i int) uint64 { return uint64(c.uintValue(i, 64, "uint64")) }
+func (c *CallCtxt) Uint64(i int) uint64 { return c.uintValue(i, 64, "uint64") }
 
 func (c *CallCtxt) uintValue(i, bits int, typ string) uint64 {
 	x := value.Make(c.ctx, c.args[i])
